python-worker/requirements.txt
# requirements.txt
grpcio==1.78.0
protobuf==6.31.1
